Dispersions.jl

This project contains

Description

Common Functions

The following functions are available for all grids.

Basic access functions

Convenience functions to access fields of k grid structs.

Dispersions.Nk — Method

Nk(kG::T) where T <: KGrid

Total number of k points (length of kGrid.kGrid for full grids).

Dispersions.gridPoints — Method

gridPoints(kG::T)::Int where T <: KGrid

Number of grid points as integer value. Nk needs to be accessible for all implemented full and reduced k grids.